Alterações em updates

+++ backend-api/RiftARENA/Models/Services/teamService.cs
@@ -137,31 +137,32 @@ namespace RiftArena.Models.Services
                 throw new AppException("Team not found!");
 
 
-            if (team.Name != teamSer.Name)
+            if (team.Name != teamSer.Name && team.Name != null)
             {
 
                 if (_context.Teams.Any(x => x.Name == team.Name))
                     throw new AppException("Team name " + team.Name + " is already taken");
+                else
+                    teamSer.Name = team.Name;
             }
 
-            if (team.Tag != teamSer.Tag)
+            if (team.Tag != teamSer.Tag && team.Tag != null)
             {
 
                 if (_context.Teams.Any(x => x.Tag == team.Tag))
                     throw new AppException("Team tag " + team.Tag + " is already taken");
+                else
+                    teamSer.Tag = team.Tag;
             }
 
-            if (team.Poster != teamSer.Poster)
+            if (team.Poster != teamSer.Poster && team.Poster != null)
             {
                 if (File.Exists(teamSer.Poster))
                 {
                     File.Delete(teamSer.Poster);
                 }
+                teamSer.Poster = team.Poster;
             }
-            teamSer.Poster = team.Poster;
-            teamSer.Name = team.Name;
-            teamSer.Tag = team.Tag;
-
 
             _context.Teams.Update(teamSer);
             _context.SaveChanges();
