+++ backend-api/RiftARENA/Models/Services/userServices.cs
@@ -242,22 +242,21 @@ namespace RiftArena.Models.Services
             if (user == null)
                 throw new AppException("User not found");
 
-            if (userParam.Nickname != user.Nickname)
-            {
-                // username has changed so check if the new username is already taken
-                if (_context.Users.Any(x => x.Nickname == userParam.Nickname))
-                    throw new AppException("Username " + userParam.Nickname + " is already taken");
-            }
-            if(userParam.Poster != user.Poster)
+            
+            if(userParam.Poster != user.Poster && userParam != null)
             {
                 if (File.Exists(user.Poster))
                 {
                     File.Delete(user.Poster);
                 }
+                user.Poster = userParam.Poster;
+
             }
-            user.Poster = userParam.Poster;
-            user.Email = userParam.Email;
-            user.Nickname = userParam.Nickname;
+            if (userParam.Email != null)
+            {
+                user.Email = userParam.Email;
+            }
+
 
             // update password if it was entered
             if (!string.IsNullOrWhiteSpace(password))
